added logo to navbar added and  aboutus with modal popup
