+++ client/src/Component/BaseLayout.js
@@ -1,6 +1,7 @@
 import React from "react";
 import Navbar from './Navbar';
-import Footer from './Footer'
+import Footer from './Footer';
+
 const BaseLayout = (props) => {
   return <>
     <Navbar />
