+++ client/src/index.css
@@ -80,3 +80,8 @@ h3{
     height: 45px;
     width: 80px;
 }
+
+.about{
+    background-color: #d4e09b;
+    grid-template-columns: repeat(2, auto[col-start]);
+}
